requirements.txt
mido
python-rtmidi
pycaw
comtypes
keyboard
pystray
pillow